Running resharper Code Cleanup

+++ nclgl/Keyboard.cpp
@@ -1,34 +1,37 @@
 #include "Keyboard.h"
 
-Keyboard::Keyboard(HWND &hwnd)	{
+Keyboard::Keyboard(HWND& hwnd)
+{
 	//Initialise the arrays to false!
-	ZeroMemory(keyStates,  KEYBOARD_MAX * sizeof(bool));
+	ZeroMemory(keyStates, KEYBOARD_MAX * sizeof(bool));
 	ZeroMemory(holdStates, KEYBOARD_MAX * sizeof(bool));
 
 	//Tedious windows RAW input stuff
-	rid.usUsagePage		= HID_USAGE_PAGE_GENERIC;		//The keyboard isn't anything fancy
-    rid.usUsage			= HID_USAGE_GENERIC_KEYBOARD;	//but it's definitely a keyboard!
-    rid.dwFlags			= RIDEV_INPUTSINK;				//Yes, we want to always receive RAW input...
-    rid.hwndTarget		= hwnd;							//Windows OS window handle
-    RegisterRawInputDevices(&rid, 1, sizeof(rid));		//We just want one keyboard, please!
+	rid.usUsagePage = HID_USAGE_PAGE_GENERIC; //The keyboard isn't anything fancy
+	rid.usUsage = HID_USAGE_GENERIC_KEYBOARD; //but it's definitely a keyboard!
+	rid.dwFlags = RIDEV_INPUTSINK; //Yes, we want to always receive RAW input...
+	rid.hwndTarget = hwnd; //Windows OS window handle
+	RegisterRawInputDevices(&rid, 1, sizeof(rid)); //We just want one keyboard, please!
 }
 
 /*
 Updates variables controlling whether a keyboard key has been
 held for multiple frames.
 */
-void Keyboard::UpdateHolds()	{
-	memcpy(holdStates,keyStates,KEYBOARD_MAX * sizeof(bool));
+void Keyboard::UpdateHolds()
+{
+	memcpy(holdStates, keyStates, KEYBOARD_MAX * sizeof(bool));
 }
 
 /*
 Sends the keyboard to sleep, so it doesn't process any
 keypresses until it receives a Wake()
 */
-void Keyboard::Sleep()	{
-	isAwake = false;	//Night night!
+void Keyboard::Sleep()
+{
+	isAwake = false; //Night night!
 	//Prevents incorrectly thinking keys have been held / pressed when waking back up
-	ZeroMemory(keyStates,  KEYBOARD_MAX * sizeof(bool));
+	ZeroMemory(keyStates, KEYBOARD_MAX * sizeof(bool));
 	ZeroMemory(holdStates, KEYBOARD_MAX * sizeof(bool));
 }
 
@@ -45,8 +48,10 @@ bool Keyboard::KeyDown(const KeyboardKeys key) const
 Returns if the key is down, and has been held down for multiple updates. 
 Doesn't need bounds checking - a KeyboardKeys enum is always in range
 */
-bool Keyboard::KeyHeld(const KeyboardKeys key)	{
-	if(KeyDown(key) && holdStates[key]) {
+bool Keyboard::KeyHeld(const KeyboardKeys key)
+{
+	if (KeyDown(key) && holdStates[key])
+	{
 		return true;
 	}
 	return false;
@@ -56,7 +61,8 @@ bool Keyboard::KeyHeld(const KeyboardKeys key)	{
 Returns true only if the key is down, but WASN't down last update.
 Doesn't need bounds checking - a KeyboardKeys enum is always in range
 */
-bool Keyboard::KeyTriggered(const KeyboardKeys key)	 {
+bool Keyboard::KeyTriggered(const KeyboardKeys key)
+{
 	return (KeyDown(key) && !KeyHeld(key));
 }
 
@@ -110,12 +116,15 @@ bool Keyboard::KeyTriggered(const std::string& mappingName)
 /*
 Updates the keyboard state with data received from the OS.
 */
-void Keyboard::Update(RAWINPUT* raw)	{
-	if(isAwake)	{
-		DWORD key = (DWORD)raw->data.keyboard.VKey;
+void Keyboard::Update(RAWINPUT* raw)
+{
+	if (isAwake)
+	{
+		DWORD key = raw->data.keyboard.VKey;
 
 		//We should do bounds checking!
-		if(key < 0 || key > KEYBOARD_MAX)	{
+		if (key < 0 || key > KEYBOARD_MAX)
+		{
 			return;
 		}
